Add support for 'css extensions'
This commit makes it possible to create a json file with so-called css
extensions. The json file contains an object with keys that equal the
classnames that will be given to the images. The corresponding values
are strings or arrays of strings. The strings are css classnames that
will get the same styling (background image).